moved client code to client package
